add CSV upload functionality for bulk holding creation, implement preview and save features, and update Holding entity for compatibility

+++ src/main/java/com/maheshbhatt/financialnuggets/entity/HoldingEntity.java
@@ -28,10 +28,10 @@ public class HoldingEntity {
     private Long quantity;
 
     @Column(nullable = false)
-    private BigDecimal marketValue;
+    private Long marketValue;
 
     @Column(nullable = false)
-    private BigDecimal percentageOfAum;
+    private Long percentageOfAum;
 
     private LocalDate reportingDate;
